+++ tests/integration/helpers/mod.rs
@@ -2,6 +2,7 @@ use crate::helpers::project::Project;
 use crate::helpers::project_builder::tempdir;
 use indoc::indoc;
 
+pub mod arg_builder;
 pub mod prelude;
 pub mod project;
 pub mod project_builder;
